output lists with each
the variable cratedContacts is an empty array that is populated via the addContact function if all fields are valid. The array createdContacts is overwritten to hold any data input per the name, title, image, and description fields. This all comes together when using each to iterate over created contacts with the input data held in the createdContacts array