created Modal component, added css to style tag, used <slot> tag to pass content into the div of the modal
used variable "closable" on App.svelte's Modal component to conditionally set functionality of the buttons of the modal (Confirm, Agree) by adding the variable "didAgree" which takes the argument "agreed"  on the slot of the Modal.svelte's footer. Using let:didAgree={closable} on the Modal component, with disabled={!closable} invoked on the slot of the footer.